refactor: move contexts to a separate folder
