+++ src/templates/docs/cloud.js
@@ -5,12 +5,12 @@ import { PageInfo } from 'components/pages/doc-welcome';
 import { CtaDoc } from 'components/shared/cta-doc';
 import { Trait } from 'components/shared/trait';
 import docPageContent from 'components/templates/doc-page/doc-page-content/doc-page-content.module.scss';
+import LocaleProvider from 'contexts/locale-provider';
 import { Link } from 'gatsby';
 import { useScrollToAnchor } from 'hooks';
 import { DocLayout } from 'layouts/doc-layout';
 import React, { useRef, useState, useEffect } from 'react';
 import { StickyContainer, Sticky } from 'react-sticky';
-import LocaleProvider from 'templates/docs/locale-provider';
 import { isInIFrame } from 'utils';
 import SeoMetadata from 'utils/seo-metadata';
 import { app } from 'utils/urls';
